using procedure component
