+++ ARSIS-Telemetry-GroundControl/groundcontrol-fe/src/App.css
@@ -10,7 +10,7 @@
 
 .App-logo {
   width: 8vw;
-  pointer-events: none;
+  pointer-events: none; 
 }
 
 .Container-primary {
