+++ ARSIS-Telemetry-GroundControl/groundcontrol-fe/src/components/procedures/CreateProcedureMenu.js
@@ -1,8 +1,8 @@
 import React, { useContext } from "react";
 import { ProcedureContext } from "../../pages/Procedures";
-import Task from "./Task"
+import Procedure from "./Procedure"
 
-const CreateProcedureMenu = (props) => {
+const CreateProcedureMenu = () => {
   const { func, tabs } = useContext(ProcedureContext)
   const [procedure, setProcedure] = func
   const [tab, setTab] = tabs
@@ -42,72 +42,13 @@ const CreateProcedureMenu = (props) => {
     setTab();
   };
 
-  const handleAddTask = () => {
-    let procedureData = { ...procedure };
-    procedureData.taskList.push({
-      name: "",
-      summary: "",
-      stepList: [{ type: "", body: "", nextTask: "" }],
-    });
-    setProcedure(procedureData);
-  };
-
-  const handleRemoveTask = () => {
-    let procedureData = { ...procedure };
-    procedureData.taskList.pop();
-    setProcedure(procedureData);
-  };
-
-  const handleAddStep = (index) => {
-    let procedureData = { ...procedure };
-    procedureData.taskList[index].stepList.push({
-      type: "",
-      body: "",
-      nextTask: "",
-    });
-    setProcedure(procedureData);
-  };
-  const handleRemoveStep = (index) => {
-    let procedureData = { ...procedure };
-    procedureData.taskList[index].stepList.pop();
-    setProcedure(procedureData);
-  };
   return (
     <div className="Container-primary">
       <form onSubmit={handleCreate}>
         <h3>Create Procedure</h3>
         <label>Procedure Details</label>
-        <input
-          type="text"
-          style={{ maxWidth: "300px" }}
-          name="proc-name"
-          placeholder="Name"
-        />
-        <textarea
-          type="text"
-          style={{ maxWidth: "300px" }}
-          name="proc-summary"
-          placeholder="Summary"
-        />
-        <label>Tasks</label>
-        <div className="Wrapped-list">
-          {procedure.taskList.length > 0 ? (
-            procedure.taskList.map((task, i) => {
-              return <Task key={i} task={task} i={i} />;
-            })
-          ) : (
-            <p>No tasks</p>
-          )}
-        </div>
-        <div className="Button-group">
-          <button type="button" onClick={handleAddTask}>
-            Add Task
-          </button>
-          <button type="button" onClick={handleRemoveTask}>
-            Remove Task
-          </button>
+          <Procedure />
           <button type="submit">Create</button>
-        </div>
       </form>
     </div>
   );
