+++ ARSIS-Telemetry-GroundControl/groundcontrol-fe/src/pages/Procedures.css
@@ -3,10 +3,19 @@
 }
 
 .Procedure {
+  display: flex;
+  flex-direction: column;
   padding: 0.8rem;
   margin: 0.5rem 1rem 1rem 0rem;
-  border-style: solid;
-  border-width: 1px;
+}
+
+.Procedure-simple {
+  display: flex;
+  flex-direction: column;
+  padding: 0.8rem;
+  margin: 0.5rem 1rem 1rem 0rem;
+  color: white;
+  background-color: rgb(0, 87, 201);
 }
 
 .Wrapped-list {
